pyproject.toml
[project]
authors = [{name = "JaeKyoung", email = "dreamfosh35@gmail.com"}]
dependencies = [
    "google-api-python-client>=2.149.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "beautifulsoup4>=4.12.3",
    "python-dotenv>=1.0.1", "slack-sdk>=3.41.0,<4", "ollama>=0.6.1,<0.7", "click>=8.3.1,<9", "httpx>=0.27,<1",
]
name = "larklab"
requires-python = ">= 3.12"
version = "0.1.0"

[project.scripts]
larklab = "larklab.main:main"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/larklab"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.pypi-dependencies]
larklab = { path = ".", editable = true }

[tool.pixi.tasks]
digest = "larklab"
